backend/pyproject.toml
[project]
name = "agentarena-backend"
version = "0.1.0"
description = "AgentArena - AI Agent Benchmark Platform Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiomysql>=0.2.0",
    "redis>=5.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "tenacity>=8.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "openpyxl>=3.1.0",
    "pymysql>=1.1.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
agentarena-api = "agentarena.main:run"
agentarena-worker = "agentarena.worker_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentarena"]
